refactor: remove copied code
