+++ x/gastracker/keeper/keeper_contract_gas_processor.go
@@ -66,7 +66,7 @@ func (k Keeper) IngestGasRecord(ctx types.Context, records []wasmTypes.ContractG
 
 		k.TrackContractGasUsage(ctx, contractAddress, wasmTypes.GasConsumptionInfo{
 			SDKGas: record.OriginalGas.SDKGas,
-			VMGas:  k.wasmGasRegister.FromWasmVMGas(record.OriginalGas.VMGas),
+			VMGas:  k.WasmGasRegister.FromWasmVMGas(record.OriginalGas.VMGas),
 		}, operation)
 	}
