+++ x/gastracker/keeper/keeper_test.go
@@ -1,117 +1,19 @@
-package keeper
+package keeper_test
 
 import (
-	"crypto/rand"
-	mintTypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	"os"
+	"github.com/archway-network/archway/x/gastracker/testutil"
 	"testing"
-	"time"
 
-	wasmKeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmTypes "github.com/CosmWasm/wasmd/x/wasm/types"
 
-	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/store"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	paramsTypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/stretchr/testify/require"
-	tmLog "github.com/tendermint/tendermint/libs/log"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	db "github.com/tendermint/tm-db"
 
 	"github.com/archway-network/archway/x/gastracker"
-	gstTypes "github.com/archway-network/archway/x/gastracker"
 )
 
-type TestContractInfoView struct {
-	adminMap     map[string]string
-	defaultAdmin string
-}
-
-func NewTestContractInfoView(defaultAdmin string) *TestContractInfoView {
-	return &TestContractInfoView{
-		adminMap:     make(map[string]string),
-		defaultAdmin: defaultAdmin,
-	}
-}
-
-func (t *TestContractInfoView) GetContractInfo(_ sdk.Context, contractAddress sdk.AccAddress) *wasmTypes.ContractInfo {
-	if admin, ok := t.adminMap[contractAddress.String()]; ok {
-		return &wasmTypes.ContractInfo{Admin: admin}
-	} else {
-		return &wasmTypes.ContractInfo{Admin: t.defaultAdmin}
-	}
-}
-
-func (t *TestContractInfoView) AddContractToAdminMapping(contractAddress string, admin string) {
-	t.adminMap[contractAddress] = admin
-}
-
-var _ ContractInfoView = &TestContractInfoView{}
-
-type subspace struct {
-	params gstTypes.Params
-}
-
-func (s *subspace) GetParamSet(_ sdk.Context, ps paramsTypes.ParamSet) {
-	*ps.(*gstTypes.Params) = s.params
-}
-
-func (s *subspace) SetParamSet(_ sdk.Context, paramset paramsTypes.ParamSet) {
-	s.params = *paramset.(*gastracker.Params)
-}
-
-type mockMinter struct{}
-
-func (t mockMinter) GetParams(_ sdk.Context) (params mintTypes.Params) {
-	return mintTypes.Params{
-		MintDenom:     "test",
-		BlocksPerYear: 100,
-	}
-}
-
-func (t mockMinter) GetMinter(_ sdk.Context) (minter mintTypes.Minter) {
-	return mintTypes.Minter{
-		AnnualProvisions: sdk.NewDec(76500),
-	}
-}
-
-func createTestBaseKeeperAndContext(t *testing.T, contractAdmin sdk.AccAddress) (sdk.Context, Keeper) {
-	memDB := db.NewMemDB()
-	ms := store.NewCommitMultiStore(memDB)
-	storeKey := sdk.NewKVStoreKey("TestStore")
-	ms.MountStoreWithDB(storeKey, sdk.StoreTypeIAVL, memDB)
-	err := ms.LoadLatestVersion()
-	require.NoError(t, err, "Loading latest version should not fail")
-	encodingConfig := simapp.MakeTestEncodingConfig()
-	appCodec := encodingConfig.Marshaler
-
-	subspace := subspace{}
-
-	keeper := Keeper{
-		key:              storeKey,
-		cdc:              appCodec,
-		paramSpace:       &subspace,
-		contractInfoView: NewTestContractInfoView(contractAdmin.String()),
-		wasmGasRegister:  wasmKeeper.NewDefaultWasmGasRegister(),
-		mintKeeper:       mockMinter{},
-	}
-
-	ctx := sdk.NewContext(ms, tmproto.Header{
-		Height: 1234567,
-		Time:   time.Date(2020, time.April, 22, 12, 0, 0, 0, time.UTC),
-	}, false, tmLog.NewTMLogger(os.Stdout))
-
-	params := gastracker.DefaultParams()
-	subspace.SetParamSet(ctx, &params)
-	return ctx, keeper
-}
-func CreateTestKeeperAndContext(t *testing.T, contractAdmin sdk.AccAddress) (sdk.Context, Keeper) {
-	return createTestBaseKeeperAndContext(t, contractAdmin)
-}
-
 func TestKeeper_Tracking(t *testing.T) {
-	ctx, k := CreateTestKeeperAndContext(t, nil)
+	ctx, k := testutil.CreateTestKeeperAndContext(t, nil)
 	k.TrackNewBlock(ctx)
 	k.TrackNewTx(ctx, sdk.NewDecCoins(sdk.NewDecCoin("test", sdk.NewInt(10))), 100_000)
 	k.TrackContractGasUsage(ctx, sdk.AccAddress("exec"), wasmTypes.GasConsumptionInfo{
@@ -133,7 +35,7 @@ func TestKeeper_Tracking(t *testing.T) {
 }
 
 func TestKeeper_GetAndIncreaseTxIdentifier_ResetTxIdentifier(t *testing.T) {
-	ctx, k := CreateTestKeeperAndContext(t, nil)
+	ctx, k := testutil.CreateTestKeeperAndContext(t, nil)
 	k.ResetTxIdentifier(ctx)
 	require.Equal(t, uint64(0), k.GetAndIncreaseTxIdentifier(ctx))
 	require.Equal(t, uint64(1), k.GetAndIncreaseTxIdentifier(ctx))
@@ -141,7 +43,7 @@ func TestKeeper_GetAndIncreaseTxIdentifier_ResetTxIdentifier(t *testing.T) {
 }
 
 func TestKeeper_UpdateGetDappInflationaryRewards(t *testing.T) {
-	ctx, k := CreateTestKeeperAndContext(t, nil)
+	ctx, k := testutil.CreateTestKeeperAndContext(t, nil)
 	t.Run("success", func(t *testing.T) {
 		rewards := k.UpdateDappInflationaryRewards(ctx, k.GetParams(ctx))
 		gotRewards, err := k.GetDappInflationaryRewards(ctx, ctx.BlockHeight())
@@ -160,10 +62,10 @@ func TestKeeper_UpdateGetDappInflationaryRewards(t *testing.T) {
 func TestContractMetadataHandling(t *testing.T) {
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	// Should return appropriate error when contract metadata is not found
 	_, err := keeper.GetContractMetadata(ctx, spareAddress[1])
 	require.EqualError(
@@ -427,10 +329,10 @@ func TestContractMetadataHandling(t *testing.T) {
 func TestCreateOrMergeLeftOverRewardEntry(t *testing.T) {
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
-	ctx, keeper := CreateTestKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 
 	_, err := keeper.GetLeftOverRewardEntry(ctx, spareAddress[1])
 	require.EqualError(t, err, gastracker.ErrRewardEntryNotFound.Error(), "Getting left over reward entry should fail")
@@ -534,10 +436,10 @@ func TestCreateOrMergeLeftOverRewardEntry(t *testing.T) {
 func TestCalculateUpdatedGas(t *testing.T) {
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
-	ctx, keeper := CreateTestKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 
 	// No change in updated gas when contract's metadata does not exists
 	gasRecord := wasmTypes.ContractGasRecord{
@@ -612,10 +514,10 @@ func TestCalculateUpdatedGas(t *testing.T) {
 func TestIngestionOfGasRecords(t *testing.T) {
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 
 	keeper.TrackNewBlock(ctx)
 
@@ -629,7 +531,7 @@ func TestIngestionOfGasRecords(t *testing.T) {
 			ContractAddress: spareAddress[1].String(),
 			OriginalGas: wasmTypes.GasConsumptionInfo{
 				SDKGas: 4,
-				VMGas:  keeper.wasmGasRegister.ToWasmVMGas(2),
+				VMGas:  keeper.WasmGasRegister.ToWasmVMGas(2),
 			},
 		},
 	})
@@ -667,7 +569,7 @@ func TestIngestionOfGasRecords(t *testing.T) {
 			ContractAddress: spareAddress[1].String(),
 			OriginalGas: wasmTypes.GasConsumptionInfo{
 				SDKGas: 2,
-				VMGas:  keeper.wasmGasRegister.ToWasmVMGas(1),
+				VMGas:  keeper.WasmGasRegister.ToWasmVMGas(1),
 			},
 		},
 		{
@@ -675,7 +577,7 @@ func TestIngestionOfGasRecords(t *testing.T) {
 			ContractAddress: spareAddress[2].String(),
 			OriginalGas: wasmTypes.GasConsumptionInfo{
 				SDKGas: 3,
-				VMGas:  keeper.wasmGasRegister.ToWasmVMGas(2),
+				VMGas:  keeper.WasmGasRegister.ToWasmVMGas(2),
 			},
 		},
 		{
@@ -683,7 +585,7 @@ func TestIngestionOfGasRecords(t *testing.T) {
 			ContractAddress: spareAddress[3].String(),
 			OriginalGas: wasmTypes.GasConsumptionInfo{
 				SDKGas: 4,
-				VMGas:  keeper.wasmGasRegister.ToWasmVMGas(3),
+				VMGas:  keeper.WasmGasRegister.ToWasmVMGas(3),
 			},
 		},
 	})
@@ -715,10 +617,10 @@ func TestIngestionOfGasRecords(t *testing.T) {
 func TestAddContractGasUsage(t *testing.T) {
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 
 	keeper.TrackNewBlock(ctx)
 
@@ -772,7 +674,7 @@ func TestAddContractGasUsage(t *testing.T) {
 // Test initialization of block tracking data for new block as well as marking end of the block for current block tracking
 // data
 func TestBlockTrackingReadWrite(t *testing.T) {
-	ctx, keeper := createTestBaseKeeperAndContext(t, sdk.AccAddress{})
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, sdk.AccAddress{})
 
 	dummyTxTracking1 := gastracker.TransactionTracking{
 		MaxGasAllowed: 500,
@@ -784,7 +686,7 @@ func TestBlockTrackingReadWrite(t *testing.T) {
 
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gastracker.BlockGasTracking{TxTrackingInfos: []gastracker.TransactionTracking{dummyTxTracking1}})
+	testutil.CreateTestBlockEntry(ctx, keeper, gastracker.BlockGasTracking{TxTrackingInfos: []gastracker.TransactionTracking{dummyTxTracking1}})
 
 	// We should be able to retrieve the block tracking info
 	currentBlockTrackingInfo := keeper.GetCurrentBlockTracking(ctx)
@@ -793,36 +695,9 @@ func TestBlockTrackingReadWrite(t *testing.T) {
 
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gastracker.BlockGasTracking{TxTrackingInfos: []gastracker.TransactionTracking{dummyTxTracking2}})
+	testutil.CreateTestBlockEntry(ctx, keeper, gastracker.BlockGasTracking{TxTrackingInfos: []gastracker.TransactionTracking{dummyTxTracking2}})
 
 	currentBlockTrackingInfo = keeper.GetCurrentBlockTracking(ctx)
 	require.Equal(t, len(currentBlockTrackingInfo.TxTrackingInfos), 1)
 	require.Equal(t, dummyTxTracking2, currentBlockTrackingInfo.TxTrackingInfos[0])
-}
-
-// TODO: this is shared test util, that is copied
-// from /module/abci_test, refactor
-func GenerateRandomAccAddress() sdk.AccAddress {
-	var address sdk.AccAddress = make([]byte, 20)
-	_, err := rand.Read(address)
-	if err != nil {
-		panic(err)
-	}
-	return address
-}
-
-// TODO: this is shared test util, that is copied
-// from /module/abci_test, refactor
-func CreateTestBlockEntry(ctx sdk.Context, k Keeper, blockTracking gstTypes.BlockGasTracking) {
-	k.TrackNewBlock(ctx)
-	for _, tx := range blockTracking.TxTrackingInfos {
-		k.TrackNewTx(ctx, tx.MaxContractRewards, tx.MaxGasAllowed)
-		for _, op := range tx.ContractTrackingInfos {
-			addr, _ := sdk.AccAddressFromBech32(op.Address)
-			k.TrackContractGasUsage(ctx, addr, wasmTypes.GasConsumptionInfo{
-				VMGas:  op.OriginalVmGas,
-				SDKGas: op.OriginalSdkGas,
-			}, op.Operation)
-		}
-	}
 }
