+++ x/gastracker/module/abci_test.go
@@ -2,30 +2,15 @@ package module
 
 import (
 	"fmt"
-	"math/rand"
-	"os"
-	"testing"
-	"time"
-
-	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/store"
-
-	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
-	wasmTypes "github.com/CosmWasm/wasmd/x/wasm/types"
+	gastracker "github.com/archway-network/archway/x/gastracker"
+	gstTypes "github.com/archway-network/archway/x/gastracker"
+	"github.com/archway-network/archway/x/gastracker/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authTypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	mintTypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	paramskeeper "github.com/cosmos/cosmos-sdk/x/params/keeper"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/abci/types"
-	tmLog "github.com/tendermint/tendermint/libs/log"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-
-	db "github.com/tendermint/tm-db"
-
-	gastracker "github.com/archway-network/archway/x/gastracker"
-	gstTypes "github.com/archway-network/archway/x/gastracker"
-	keeper "github.com/archway-network/archway/x/gastracker/keeper"
+	"testing"
 )
 
 type Behaviour int
@@ -150,7 +135,7 @@ func disableGasRebateToUser(params gstTypes.Params) gstTypes.Params {
 
 // Test the conditions under which BeginBlocker and EndBlocker should panic or not panic
 func TestABCIPanicBehaviour(t *testing.T) {
-	ctx, keeper := CreateTestKeeperAndContext(t, sdk.AccAddress{})
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, sdk.AccAddress{})
 
 	config := sdk.GetConfig()
 	config.SetBech32PrefixForAccount("archway", "archway")
@@ -175,12 +160,12 @@ func TestABCIContractMetadataCommit(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	testRewardKeeper := &TestRewardTransferKeeper{B: Log}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 
 	contractMetadatas := []gstTypes.ContractInstanceMetadata{
 		{
@@ -270,7 +255,7 @@ func TestRewardCalculation(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -295,7 +280,7 @@ func TestRewardCalculation(t *testing.T) {
 		},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -347,7 +332,7 @@ func TestRewardCalculation(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -424,7 +409,7 @@ func TestContractRewardsWithoutContractPremium(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -449,7 +434,7 @@ func TestContractRewardsWithoutContractPremium(t *testing.T) {
 		},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 	keeper.SetParams(ctx, params)
 
@@ -500,7 +485,7 @@ func TestContractRewardsWithoutContractPremium(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -573,7 +558,7 @@ func TestContractRewardsWithoutDappInflation(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -598,7 +583,7 @@ func TestContractRewardsWithoutDappInflation(t *testing.T) {
 		},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -650,7 +635,7 @@ func TestContractRewardsWithoutDappInflation(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -723,7 +708,7 @@ func TestContractRewardsWithoutGasRebate(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -746,7 +731,7 @@ func TestContractRewardsWithoutGasRebate(t *testing.T) {
 		},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -798,7 +783,7 @@ func TestContractRewardsWithoutGasRebate(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -872,7 +857,7 @@ func TestContractRewardWithoutGasRebateAndDappInflation(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -887,7 +872,7 @@ func TestContractRewardWithoutGasRebateAndDappInflation(t *testing.T) {
 		logs:     []*RewardTransferKeeperCallLogs{},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -939,7 +924,7 @@ func TestContractRewardWithoutGasRebateAndDappInflation(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -1005,7 +990,7 @@ func TestContractRewardsWithoutGasTracking(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -1020,7 +1005,7 @@ func TestContractRewardsWithoutGasTracking(t *testing.T) {
 		logs:     []*RewardTransferKeeperCallLogs{},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -1072,7 +1057,7 @@ func TestContractRewardsWithoutGasTracking(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -1139,7 +1124,7 @@ func TestContractRewardsWithoutGasRebateToUser(t *testing.T) {
 
 	var spareAddress = make([]sdk.AccAddress, 10)
 	for i := 0; i < 10; i++ {
-		spareAddress[i] = GenerateRandomAccAddress()
+		spareAddress[i] = testutil.GenerateRandomAccAddress()
 	}
 
 	type expect struct {
@@ -1164,7 +1149,7 @@ func TestContractRewardsWithoutGasRebateToUser(t *testing.T) {
 		},
 	}
 
-	ctx, keeper := createTestBaseKeeperAndContext(t, spareAddress[0])
+	ctx, keeper := testutil.CreateTestKeeperAndContext(t, spareAddress[0])
 	ctx = ctx.WithBlockGasMeter(sdk.NewGasMeter(200000))
 
 	keeper.SetParams(ctx, params)
@@ -1216,7 +1201,7 @@ func TestContractRewardsWithoutGasRebateToUser(t *testing.T) {
 	// Tracking new block with multiple tx tracking obj
 	keeper.TrackNewBlock(ctx)
 
-	CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
+	testutil.CreateTestBlockEntry(ctx, keeper, gstTypes.BlockGasTracking{TxTrackingInfos: []gstTypes.TransactionTracking{
 		{
 			MaxGasAllowed:      10,
 			MaxContractRewards: []sdk.DecCoin{firstTxMaxContractReward[0], firstTxMaxContractReward[1]},
@@ -1282,100 +1267,4 @@ func TestContractRewardsWithoutGasRebateToUser(t *testing.T) {
 	for i := 0; i < len(expected.rewardsB); i++ {
 		require.Equal(t, expected.rewardsB[i], leftOverEntry.ContractRewards[i])
 	}
-}
-
-// TODO: this is shared test util, that is copied
-// from /keeper/keeper_test, refactor
-func createTestBaseKeeperAndContext(t *testing.T, contractAdmin sdk.AccAddress) (sdk.Context, keeper.Keeper) {
-	encodingConfig := simapp.MakeTestEncodingConfig()
-	appCodec := encodingConfig.Marshaler
-
-	memDB := db.NewMemDB()
-	ms := store.NewCommitMultiStore(memDB)
-
-	mkey := sdk.NewKVStoreKey("test")
-	tkey := sdk.NewTransientStoreKey("transient_test")
-	tstoreKey := sdk.NewTransientStoreKey(gastracker.TStoreKey)
-
-	ms.MountStoreWithDB(mkey, sdk.StoreTypeIAVL, memDB)
-	ms.MountStoreWithDB(tkey, sdk.StoreTypeIAVL, memDB)
-	ms.MountStoreWithDB(storeKey, sdk.StoreTypeIAVL, memDB)
-	ms.MountStoreWithDB(tstoreKey, sdk.StoreTypeTransient, memDB)
-
-	err := ms.LoadLatestVersion()
-	require.NoError(t, err, "Loading latest version should not fail")
-
-	pkeeper := paramskeeper.NewKeeper(appCodec, encodingConfig.Amino, mkey, tkey)
-	subspace := pkeeper.Subspace(gstTypes.ModuleName)
-
-	keeper := keeper.NewGasTrackingKeeper(
-		storeKey,
-		appCodec,
-		subspace,
-		NewTestContractInfoView(contractAdmin.String()),
-		wasmkeeper.NewDefaultWasmGasRegister(),
-		&TestMintParamsKeeper{B: Log},
-	)
-
-	ctx := sdk.NewContext(ms, tmproto.Header{
-		Time: time.Date(2020, time.April, 22, 12, 0, 0, 0, time.UTC),
-	}, false, tmLog.NewTMLogger(os.Stdout))
-
-	params := gstTypes.DefaultParams()
-	subspace.SetParamSet(ctx, &params)
-	return ctx, keeper
-}
-
-func CreateTestKeeperAndContext(t *testing.T, contractAdmin sdk.AccAddress) (sdk.Context, keeper.Keeper) {
-	return createTestBaseKeeperAndContext(t, contractAdmin)
-}
-
-func CreateTestBlockEntry(ctx sdk.Context, k keeper.Keeper, blockTracking gstTypes.BlockGasTracking) {
-	k.TrackNewBlock(ctx)
-	for _, tx := range blockTracking.TxTrackingInfos {
-		k.TrackNewTx(ctx, tx.MaxContractRewards, tx.MaxGasAllowed)
-		for _, op := range tx.ContractTrackingInfos {
-			addr, _ := sdk.AccAddressFromBech32(op.Address)
-			k.TrackContractGasUsage(ctx, addr, wasmTypes.GasConsumptionInfo{
-				VMGas:  op.OriginalVmGas,
-				SDKGas: op.OriginalSdkGas,
-			}, op.Operation)
-		}
-	}
-}
-
-type TestContractInfoView struct {
-	keeper.ContractInfoView
-	adminMap     map[string]string
-	defaultAdmin string
-}
-
-func NewTestContractInfoView(defaultAdmin string) *TestContractInfoView {
-	return &TestContractInfoView{
-		adminMap:     make(map[string]string),
-		defaultAdmin: defaultAdmin,
-	}
-}
-
-func (t *TestContractInfoView) GetContractInfo(_ sdk.Context, contractAddress sdk.AccAddress) *wasmTypes.ContractInfo {
-	if admin, ok := t.adminMap[contractAddress.String()]; ok {
-		return &wasmTypes.ContractInfo{Admin: admin}
-	} else {
-		return &wasmTypes.ContractInfo{Admin: t.defaultAdmin}
-	}
-}
-
-func (t *TestContractInfoView) AddContractToAdminMapping(contractAddress string, admin string) {
-	t.adminMap[contractAddress] = admin
-}
-
-var _ keeper.ContractInfoView = &TestContractInfoView{}
-
-func GenerateRandomAccAddress() sdk.AccAddress {
-	var address sdk.AccAddress = make([]byte, 20)
-	_, err := rand.Read(address)
-	if err != nil {
-		panic(err)
-	}
-	return address
 }
